+++ checkpoint/orbax/checkpoint/checkpoint_manager.py
@@ -29,11 +29,14 @@ from jax.experimental import multihost_utils
 from jax.experimental.array_serialization import serialization
 import numpy as np
 from orbax.checkpoint import utils
+# pylint: disable=g-importing-member
+from orbax.checkpoint.abstract_checkpoint_manager import AbstractCheckpointManager
 from orbax.checkpoint.abstract_checkpointer import AbstractCheckpointer
 from orbax.checkpoint.async_checkpointer import AsyncCheckpointer
 from orbax.checkpoint.checkpointer import Checkpointer
 from orbax.checkpoint.json_checkpoint_handler import JsonCheckpointHandler
 from orbax.checkpoint.proto_checkpoint_handler import ProtoCheckpointHandler
+# pylint: enable=g-importing-member
 PyTree = Any
 CheckpointDirs = Tuple[str, str]
 SaveParams = Mapping[str, Any]
@@ -179,25 +182,8 @@ class CheckpointInfo:
     return self.step == other.step and self.time == other.time
 
 
-class CheckpointManager:
-  """A generic, synchronous CheckpointManager implementation.
-
-  Allows a user to save and restore objects for which a Checkpointer
-  implementation exists (e.g. PyTreeCheckpointer for PyTrees). The class
-  keeps track of multiple checkpointable objects in the following structure::
-
-    path/to/directory/    (top-level directory)
-      0/    (step)
-        params/    (first saveable)
-          ...
-        metadata/    (second saveable)
-          ...
-      1/    (step)
-        ...
-      2/    (step)
-        ...
-      ...
-  """
+class CheckpointManager(AbstractCheckpointManager):
+  """A generic, synchronous AbstractCheckpointManager implementation."""
 
   def __init__(
       self,
@@ -248,12 +234,12 @@ class CheckpointManager:
     elif isinstance(checkpointers, dict):
       for item in [k for k in checkpointers if k in RESERVED_ITEM_NAMES]:
         raise ValueError(
-            f'Found {item} in `checkpointers`; this is a reserved'
-            ' key.'
+            f'Found {item} in `checkpointers`; this is a reserved key.'
         )
     else:
       raise ValueError(
-          f'Invalid type for `checkpointers`. Found {checkpointers}.')
+          f'Invalid type for `checkpointers`. Found {checkpointers}.'
+      )
 
     self._checkpointers = checkpointers
     self._options = options or CheckpointManagerOptions()
@@ -292,18 +278,11 @@ class CheckpointManager:
 
   @property
   def directory(self) -> epath.Path:
+    """See superclass documentation."""
     return self._directory
 
   def all_steps(self, read: bool = False) -> Sequence[int]:
-    """Returns all steps tracked by the manager.
-
-    Args:
-      read: If True, forces a read directly from the storage location.
-        Otherwise, a cached result can be returned.
-
-    Returns:
-      A sequence of steps (integers)
-    """
+    """See superclass documentation."""
     if read:
       if self._options.single_host_load_and_broadcast:
         max_steps = len(list(self.directory.iterdir()))
@@ -324,24 +303,12 @@ class CheckpointManager:
       return [ckpt.step for ckpt in self._checkpoints]
 
   def latest_step(self) -> Optional[int]:
-    """Returns the latest step saved.
-
-    Returns None if no steps have been saved.
-
-    Returns:
-      A step (int) or None if no steps are present.
-    """
+    """See superclass documentation."""
     steps = self.all_steps(read=False)
     return max(steps) if steps else None
 
   def best_step(self) -> Optional[int]:
-    """Returns the best step saved, as defined by `options.best_fn`.
-
-    Returns None if no steps have been saved.
-
-    Returns:
-      A step (int) or None if no steps are present.
-    """
+    """See superclass documentation."""
     if not self._track_best:
       return self.latest_step()
     if not self._checkpoints:
@@ -352,20 +319,11 @@ class CheckpointManager:
     return sorted_checkpoints[-1].step
 
   def reached_preemption(self, step: int) -> bool:
-    """Returns True if a preemption sync point has been reached."""
+    """See superclass documentation."""
     return utils.reached_preemption(step)
 
   def should_save(self, step: int) -> bool:
-    """Returns True if a checkpoint should be saved for the current step.
-
-    This depends the previous step and save interval.
-
-    Args:
-      step: int
-
-    Returns:
-      True if the checkpoint should be saved.
-    """
+    """See superclass documentation."""
     if self.reached_preemption(step):
       return True
     last_checkpoint_step = (
@@ -403,7 +361,7 @@ class CheckpointManager:
     return utils.create_tmp_directory(directory)
 
   def delete(self, step: int):
-    """Deletes a step checkpoint."""
+    """See superclass documentation."""
     if step not in self.all_steps():
       raise ValueError(f'Requested deleting a non-existent step: {step}.')
     self._delete_directory(step)
@@ -419,69 +377,7 @@ class CheckpointManager:
                                                            SaveParams]]] = None,
            metrics: Optional[PyTree] = None,
            force: Optional[bool] = False) -> bool:
-    """Saves the provided items.
-
-    This method should be called by all hosts - process synchronization and
-    actions that need to be performed on only one host are managed internally.
-
-    Items and save_kwargs must have a top-level structure matching that of
-    self._checkpointers, meaning that for every key in items and save_kwargs, a
-    corresponding key must be present in self._checkpointers.
-
-    Items takes a form similar to the following::
-
-      {
-        'params': PyTree(),
-        'metadata': <nested k/v>,
-        ...
-      }
-      Similarly, save_kwargs takes the form:
-      {
-        'params': {
-          <kwargs for PyTreeCheckpointHandler.save>
-        },
-        'metadata': {
-          <kwargs for JsonCheckpointHandler.save>
-        }
-        ...
-      }
-
-    The kwargs under 'params' correspond to PyTreeCheckpointHandler.save. If a
-    key is not present in save_kwargs, it is assumed that no kwargs are needed
-    for saving that item. If not provided at all, it is assumed that no items
-    need extra kwargs for saving.
-
-    Note that if a single Checkpointer was provided at construction time,
-    `items` must be a singular saveable object, and `save_kwargs` must be the
-    kwargs needed by a single Checkpointer.
-
-    Args:
-      step: current step, int
-      items: a savable object, or a dictionary of object name to savable object.
-      save_kwargs: save kwargs for a single Checkpointer, or a dictionary of
-        object name to kwargs needed by the Checkpointer implementation to save
-        the object.
-      metrics: a dictionary of metric name (string) to numeric value to be
-        tracked along with this checkpoint. Required if `options.best_fn` is
-        set. Allows users to specify a metric value to determine which
-        checkpoints are best and should be kept (in conjunction with
-        `options.max_to_keep`).
-      force: if `True`, this method will attempt to save a checkpoint
-        regardless of the result of `CheckpointManager.should_save(step)`. By
-        default, `save` will only write a checkpoint to disk when the options
-        permit, e.g. when `step` is in `options.save_interval_steps` or
-        `options.save_on_steps`.
-        Setting `force=True` will not overwrite existing checkpoints.
-
-    Returns:
-      bool indicating whether a save operation was performed.
-    Raises:
-      ValueError: if `track_best` was indicated but `metrics` is not provided.
-      ValueError: directory creation failed.
-      ValueError: if an item is provided for which no `Checkpointer` is
-      found.
-      ValueError: if the checkpoint already exists.
-    """
+    """See superclass documentation."""
     if not force and not self.should_save(step):
       return False
     if self.reached_preemption(step):
@@ -547,69 +443,7 @@ class CheckpointManager:
                                      Mapping[str, RestoreParams]]] = None,
       directory: Optional[epath.PathLike] = None
   ) -> Union[Any, Mapping[str, Any]]:
-    """Restores from the given step and provided items.
-
-    This method should be called by all hosts - process synchronization and
-    actions that need to be performed on only one host are managed internally.
-
-    Items and restore_kwargs must have a top-level structure matching that of
-    self._checkpointers, meaning that for every key in items and restore_kwargs,
-    a
-    corresponding key must be present in self._checkpointers.
-
-    Items takes a form similar to the following::
-
-      {
-        'params': PyTree(),
-        'metadata': <nested k/v>,
-        ...
-      }
-
-    Items may not be provided at all, in which case it the items restored are
-    those specified in self._checkpointers, and item=None is provided to
-    Checkpointer.restore. Similarly, an item may be omitted from `items`,
-    in
-    which case item=None will be provided to Checkpointer.restore.
-
-    Similarly, restore_kwargs takes the form::
-
-      {
-        'params': {
-          'meshes': PyTree(),
-          'mesh_axes': PyTree(),
-        },
-        'metadata': {
-          <kwargs for JsonCheckpointHandler.save>
-        }
-        ...
-      }
-
-    The kwargs under 'params' correspond to PyTreeCheckpointHandler.restore. If
-    a key is not present in restore_kwargs, it is assumed that no kwargs are
-    needed for restoring that item. If not provided at all, it is assumed that
-    no items need extra kwargs for restoring.
-
-    Note that if a single Checkpointer was provided at construction time,
-    `items` must be a singular saveable object, and `restore_kwargs` must be the
-    kwargs needed by a single Checkpointer.
-
-    Args:
-      step: current step, int
-      items: a restoreable object, or a dictionary of object name to restorable
-        object.
-      restore_kwargs: restore kwargs for a single Checkpointer, or a dictionary
-        of object name to kwargs needed by the Checkpointer implementation to
-        restore the object.
-      directory: if provided, uses the given directory rather than the
-        `directory` property of this class. Can be used to restore checkpoints
-        from an independent location.
-
-    Returns:
-      A dictionary matching the structure of self._checkpointers, with one
-      object returned for each Checkpointer, or a single restored object,
-      if a
-      single item is being tracked by this manager.
-    """
+    """See superclass documentation."""
     if items is None:
       items = {}
     elif self._single_item:
@@ -657,20 +491,7 @@ class CheckpointManager:
     return restored
 
   def item_metadata(self, step: int) -> Union[Any, Mapping[str, Optional[Any]]]:
-    """For all Checkpointers, returns any metadata associated with the item.
-
-    Calls the `metadata` method for each Checkpointer and returns a
-    mapping of each item name to the restored metadata. If the manager only
-    manages a single item, a single metadata will be returned instead.
-
-    Metadata may be None for an individual item.
-
-    Args:
-      step: Step for which to retrieve metadata.
-
-    Returns:
-      A dictionary mapping name to item metadata, or a single item metadata.
-    """
+    """See superclass documentation."""
     result = {}
     for name, checkpointer in self._checkpointers.items():
       path = self._get_save_directory(step, self.directory, name)
@@ -760,7 +581,7 @@ class CheckpointManager:
       checkpointer.save(path, metadata)
 
   def metadata(self) -> Mapping[str, Any]:
-    """Returns CheckpointManager level metadata if present, empty otherwise."""
+    """See superclass documentation."""
     if self._metadata is None:
       path = self.directory / METADATA_ITEM_NAME
       if path.exists():
@@ -922,19 +743,7 @@ class CheckpointManager:
       self._checkpoints = kept_checkpoints
 
   def wait_until_finished(self, join_finalize_thread=True):
-    """Blocks until any incomplete save operations are completed.
-
-    Note that this method will typically be a no-op if all checkpointers are
-    synchronous, since old checkpoints are already cleaned up immediately after
-    completing `save`, and there is no background thread to wait for.
-
-    If some checkpointers are of type AsyncCheckpointer, however, this method
-    will wait until each of these checkpointers is finished.
-
-    Args:
-      join_finalize_thread: Whether to join the _finalize_thread. This should
-        always be True for external callers.
-    """
+    """See superclass documentation."""
     for checkpointer in self._checkpointers.values():
       if is_async_checkpointer(checkpointer):
         checkpointer.wait_until_finished()  # pytype: disable=attribute-error
@@ -948,10 +757,7 @@ class CheckpointManager:
         utils.sync_global_devices('CheckpointManager:join_finalize_thread')
 
   def check_for_errors(self):
-    """Checks for any outstanding errors in completed asynchronous save operations.
-
-    Delegates to underlying Checkpointer.
-    """
+    """See superclass documentation."""
     for checkpointer in self._checkpointers.values():
       if is_async_checkpointer(checkpointer):
         checkpointer.check_for_errors()  # pytype: disable=attribute-error
@@ -1011,7 +817,7 @@ class CheckpointManager:
       self._delete_directory(step)
 
   def close(self):
-    """Waits for outstanding operations to finish and closes Checkpointers."""
+    """See superclass documentation."""
     self.wait_until_finished()
     for c in self._checkpointers.values():
       c.close()